# doc/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://43.128.109.91:8801",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/collect/messages": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u804a\u5929\u8bb0\u5f55 (WS code=17)",
                "description": "\u5165\u53c2\u53ef\u662f {code:17,data:{list:[...]}} \u6216 {list:[...]} \u6216\u76f4\u63a5 [...] \u6570\u7ec4",
                "operationId": "messages",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collect/friends": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u8d26\u53f7\u7684\u597d\u53cb\u5217\u8868 (WS code=7)",
                "description": "\u5165\u53c2\u540c\u4e0a, \u6bcf\u6761\u81ea\u5e26 accountId \u8bf4\u660e\u5f52\u5c5e\u54ea\u4e2a\u8d26\u53f7",
                "operationId": "friends",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collect/accounts": {
            "post": {
                "tags": [
                    "\u91c7\u96c6\u63a5\u53e3 (Collect)"
                ],
                "summary": "\u91c7\u96c6\u8d26\u53f7\u5217\u8868 (WS code=1)",
                "description": "\u5165\u53c2\u53ef\u76f4\u63a5\u662f WS \u8fd4\u56de\u4f53 {code:1,data:[...]} \u6216\u8005\u53ea\u6709 data \u6570\u7ec4",
                "operationId": "accounts",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/messages": {
            "get": {
                "tags": [
                    "\u7ba1\u7406\u540e\u53f0\u67e5\u8be2 (Query)"
                ],
                "summary": "\u5206\u9875\u67e5\u8be2\u804a\u5929\u8bb0\u5f55 (\u53ef\u6309\u8d26\u53f7+\u597d\u53cb\u8fc7\u6ee4)",
                "operationId": "listMessages",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 20
                        }
                    },
                    {
                        "name": "ownerZaloId",
                        "in": "query",
                        "description": "\u5f52\u5c5e\u8d26\u53f7 (owner_zalo_id)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "peerUserId",
                        "in": "query",
                        "description": "\u5bf9\u7aef userId (fid)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "query",
                        "description": "\u7fa4 ID, 0 \u8868\u793a\u4e00\u5bf9\u4e00",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "msgType",
                        "in": "query",
                        "description": "\u6d88\u606f\u7c7b\u578b: 1=\u6587\u672c 2=\u56fe\u7247 3=\u8bed\u97f3 4=\u89c6\u9891 5=\u7cfb\u7edf 7=\u8d34\u56fe 8=\u5361\u7247",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "\u65b9\u5411: 0=\u6536 \u5176\u4ed6=\u53d1",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "contentLike",
                        "in": "query",
                        "description": "\u5185\u5bb9\u5305\u542b",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startTime",
                        "in": "query",
                        "description": "\u8d77\u59cb\u65f6\u95f4 yyyy-MM-dd HH:mm:ss",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "endTime",
                        "in": "query",
                        "description": "\u7ed3\u675f\u65f6\u95f4 yyyy-MM-dd HH:mm:ss",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RIPageZaloMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/friends": {
            "get": {
                "tags": [
                    "\u7ba1\u7406\u540e\u53f0\u67e5\u8be2 (Query)"
                ],
                "summary": "\u5206\u9875\u67e5\u8be2\u597d\u53cb\u5217\u8868 (\u53ef\u6309\u8d26\u53f7\u8fc7\u6ee4)",
                "operationId": "listFriends",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 20
                        }
                    },
                    {
                        "name": "ownerZaloId",
                        "in": "query",
                        "description": "\u5f52\u5c5e\u8d26\u53f7 (zalo_account.zalo_id)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "friendUserId",
                        "in": "query",
                        "description": "\u597d\u53cb userId (\u7cbe\u786e)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "displayName",
                        "in": "query",
                        "description": "\u597d\u53cb\u663e\u793a\u540d (\u6a21\u7cca)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "\u597d\u53cb\u624b\u673a\u53f7 (\u6a21\u7cca)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fstatus",
                        "in": "query",
                        "description": "\u5173\u7cfb: 0=\u964c\u751f\u4eba 3=\u597d\u53cb 5=\u975e\u597d\u53cb",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "deptId",
                        "in": "query",
                        "description": "\u90e8\u95e8 ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RIPageZaloFriend"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/accounts": {
            "get": {
                "tags": [
                    "\u7ba1\u7406\u540e\u53f0\u67e5\u8be2 (Query)"
                ],
                "summary": "\u5206\u9875\u67e5\u8be2 Zalo \u8d26\u53f7",
                "operationId": "listAccounts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u9875\u7801, \u4ece 1 \u5f00\u59cb",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "\u6bcf\u9875\u5927\u5c0f",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 20
                        }
                    },
                    {
                        "name": "zaloId",
                        "in": "query",
                        "description": "Zalo \u8d26\u53f7 ID (\u7cbe\u786e)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "account",
                        "in": "query",
                        "description": "\u8d26\u53f7\u624b\u673a\u53f7 (\u6a21\u7cca)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nickName",
                        "in": "query",
                        "description": "\u6635\u79f0 (\u6a21\u7cca)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "deptId",
                        "in": "query",
                        "description": "\u90e8\u95e8 ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "online",
                        "in": "query",
                        "description": "\u5728\u7ebf\u72b6\u6001: 1=\u5728\u7ebf 2=\u79bb\u7ebf",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "accountStatus",
                        "in": "query",
                        "description": "\u8d26\u53f7\u72b6\u6001",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RIPageZaloAccount"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/accounts/{id}": {
            "get": {
                "tags": [
                    "\u7ba1\u7406\u540e\u53f0\u67e5\u8be2 (Query)"
                ],
                "summary": "\u8d26\u53f7\u8be6\u60c5",
                "operationId": "accountDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RZaloAccount"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "JsonNode": {
                "type": "object"
            },
            "RMapStringObject": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
            "IPageZaloMessage": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "records": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZaloMessage"
                        }
                    },
                    "current": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pages": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "RIPageZaloMessage": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/IPageZaloMessage"
                    }
                }
            },
...
            "IPageZaloFriend": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "records": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZaloFriend"
                        }
                    },
                    "current": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pages": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "RIPageZaloFriend": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/IPageZaloFriend"
                    }
                }
            },
...
            "IPageZaloAccount": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "records": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZaloAccount"
                        }
                    },
                    "current": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pages": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "RIPageZaloAccount": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/IPageZaloAccount"
                    }
                }
            },
            "ZaloAccount": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "zaloId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "account": {
                        "type": "string"
                    },
                    "nickName": {
                        "type": "string"
                    },
                    "region": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "sex": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "headimg": {
                        "type": "string"
                    },
                    "online": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "accountStatus": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "proxyIp": {
                        "type": "string"
                    },
                    "proxyId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "friendsSum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "groupSum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msgSum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "groupId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "groupName": {
                        "type": "string"
                    },
                    "deptId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "bindId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sdbo": {
                        "type": "string"
                    },
                    "wxid": {
                        "type": "string"
                    },
                    "isShop": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "realName": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createBy": {
                        "type": "string"
                    },
                    "gmtCreate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "gmtUpdate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "collectedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "rawJson": {
                        "type": "string"
                    },
                    "deleted": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "RZaloAccount": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ZaloAccount"
                    }
                }
            }
        }
    }
}